backend/requirements/local.txt
-r base.txt

django-debug-toolbar==3.2.2
